refactor: 優先度の高いファイルをリファクタリング
